#276 Extract FakeType so ReportGenerator can exclude it.
Move FakeType out of AttributeDataExtensions into its own type, update classfilters, and bump to 10.2.0-alpha.13 so method coverage reaches 100%.
