+++ MappaVersion.targets
@@ -1,5 +1,5 @@
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
     <PropertyGroup>
-        <MappaVersion>10.2.0-alpha.12</MappaVersion>
+        <MappaVersion>10.2.0-alpha.13</MappaVersion>
     </PropertyGroup>
 </Project>
